+++ CHANGELOG.md
@@ -35,6 +35,7 @@ Expected: September 2026
 
 ### Corrected Bugs
 
+* Fixed: [XML/YANG binding uses wrong sibling namespace for same local node names](https://github.com/clicon/clixon/issues/670)
 * Fixed: [YANG incremental validation of MUST failed when an entire top-level list entry is deleted](https://github.com/clicon/clixon/issues/681)
 * Fixed: [RESTCONF PATCH/PUT requests for nested list entries stopped working](https://github.com/clicon/clixon/issues/679)